Score added, game resets on collision
Score is now being taken and shows in the title bar of the window. The
game also resets and everyone goes back to their starting positions on
the game start. Need to fix collision detection. If a player moves into
the block behind the other player, it won't detect a collision
immediately and the player goes through the block in the trail. Might be
cool to keep it though. Also, the first block in the front of the player
disappears after the game resets for the first time. Collision seems to
still be applied to that space though? Who nose.