Fixed menu, fixed way controls work, started random AI
I fixed the way the menu works, so that it actually sends meaningful
data to the main module. This is used mainly just for game mode
selection. I also changed the controls up a bit to make room for the
menu and game mode changes. Inputmaps are now global to make programming
things a little easier (and so I didn't have to return an inputmap into
a variable every time I wanted to call the playercontroller methods).
Things make more sense and are a lot cleaner. Also started setting
controls up for the random moving AI. Need to add checks for walls and
what not still.