Finished Random AI and Ordered Selection AI
Random AI is pretty bad, but it's random so I don't expect much. Ordered
Selection AI is really cool, especially when you change the order that
it's in. Might take order as a parameter from the menu if these things
are chosen. Also fixed the code to check if there was going to be a
collision soon and return possible moves in the posMoves function.