Added main file & skeleton for wallhug AI, Allow to change players without closing the game
I added a call from tronMain to gameMenu to allow you to change AI
without having to close the game. This caused errors with gameMenu. I
added an "if __name__ = __main__" thing in gameMenu, and added a file
main.py that imports gameMenu. This satisfied the error, and everything
works great.